+++ Frontend/src/lib/icons/Info.svelte
@@ -1,8 +1,30 @@
 <script lang="ts">
-  let {className = '!h-12 !w-12 mr-4 !fill-info-500'}: {className?: string | undefined} = $props();
+  interface Props {
+    className?: string | undefined;
+    fill?: boolean;
+  }
+
+  let {fill = false, className = '!h-12 !w-12 mr-4 stroke-info-500'}: Props = $props();
 </script>
 
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" class={className + ' small-icon'}>
-  <path
-    d="M2.93 17.07A10 10 0 1 1 17.07 2.93 10 10 0 0 1 2.93 17.07zm12.73-1.41A8 8 0 1 0 4.34 4.34a8 8 0 0 0 11.32 11.32zM9 5h2v6H9V5zm0 8h2v2H9v-2z" />
-</svg>
+{#if fill}
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    stroke-width="2"
+    class={className + ' small-icon fill-current !stroke-none'}>
+    <path stroke="none" d="M0 0h24v24H0z" fill="none" /><path
+      d="M17 3.34a10 10 0 1 1 -15 8.66l.005 -.324a10 10 0 0 1 14.995 -8.336m-5 11.66a1 1 0 0 0 -1 1v.01a1 1 0 0 0 2 0v-.01a1 1 0 0 0 -1 -1m0 -7a1 1 0 0 0 -1 1v4a1 1 0 0 0 2 0v-4a1 1 0 0 0 -1 -1" />
+  </svg>
+{:else}
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+    class={className + ' small-icon fill-none stroke-current'}>
+    <path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M12 12m-9 0a9 9 0 1 0 18 0a9 9 0 1 0 -18 0" /><path
+      d="M12 9v4" /><path d="M12 16v.01" />
+  </svg>
+{/if}