+++ Frontend/src/lib/index.ts
@@ -1,19 +1,9 @@
-import {notification, showModal, pricingModel, fetching} from '$store';
+import {notification, showModal, fetching} from '$store';
 import type {Cookies} from '@sveltejs/kit';
 import type {Notification} from '$type';
-import {allPricingModel} from '$type';
 import {dev} from '$app/environment';
 import {get} from 'svelte/store';
 import {token} from '$store';
-
-export function changePricingModel(model: string) {
-  const chosenModel = model.toLowerCase() as keyof typeof allPricingModel;
-  pricingModel.set({name: model, ...allPricingModel[chosenModel]});
-
-  const margin = model === 'Monthly' ? '0%' : model === 'Annually' ? '33%' : '66%';
-  const element = document.querySelector('#select-model-box') as HTMLElement;
-  element.style.left = margin;
-}
 
 export function notify(message: string, type: Notification['type'] = 'info') {
   const id = Math.floor(Math.random() * 10000);