+++ Frontend/src/lib/stores.ts
@@ -1,6 +1,6 @@
 import type {Notification, PricingModel} from '$type';
 import {writable, type Writable} from 'svelte/store';
-import {allPricingModel} from '$type';
+import {allPricingModels} from '$type';
 
 export const user: Writable<string> = writable('');
 export const token: Writable<string> = writable('');
@@ -14,8 +14,9 @@ export const scrollYProgress: Writable<number> = writable(0);
 
 export const showModal: Writable<number> = writable(0);
 export const fetching: Writable<number> = writable(0);
+
 export const selectionMenuOpen: Writable<boolean> = writable(false);
 export const selectionInputOpen: Writable<boolean> = writable(false);
 
 export const notification: Writable<Notification> = writable({id: null, message: '', type: 'info'});
-export const pricingModel: Writable<PricingModel> = writable({name: 'Monthly', ...allPricingModel.monthly});
+export const pricingModel: Writable<PricingModel> = writable({name: 'Monthly', ...allPricingModels.monthly});