+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.08] - 2024-1-06
+
+### Added
+
+- Add Complete Email Process
+- Add Email as a signup requirement
+- Add forgot password on login
+- Add Email verification
+- Add Email to settings
+- Add Tooltip component
+
 ## [0.3.07] - 2024-12-22
 
 ### Added